requirements.txt
flask==2.3.3
openai>=1.12.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytz==2024.1
psycopg2-binary==2.9.9
sentry-sdk[flask]==1.40.0
sqlalchemy